fix doctests

+++ qmcpy/integrand/asian_option.py
@@ -21,7 +21,7 @@ class AsianOption(Integrand):
         dim_frac        0
     >>> x = ac.discrete_distrib.gen_samples(2**12)
     >>> y = ac.f(x)
-    >>> y.mean()
+    >>> float(y.mean())
     1.768...
     >>> level_dims = [2,4,8]
     >>> ac2_multilevel = AsianOption(DigitalNetB2(seed=7),multilevel_dims=level_dims)
@@ -32,7 +32,7 @@ class AsianOption(Integrand):
     ...     x = ac2_single_level.discrete_distrib.gen_samples(2**12)
     ...     level_est = ac2_single_level.f(x).mean()
     ...     yml += level_est
-    >>> yml
+    >>> float(yml)
     1.779...
     """
