+++ qmcpy/integrand/hartmann6d.py
@@ -8,7 +8,7 @@ class Hartmann6d(Integrand):
     >>> h6d = Hartmann6d(DigitalNetB2(6,seed=7))
     >>> x = h6d.discrete_distrib.gen_samples(2**10)
     >>> y = h6d.f(x)
-    >>> y.mean()
+    >>> float(y.mean())
     -0.2613140309713834
     >>> h6d.true_measure
     Uniform (TrueMeasure Object)
