+++ qmcpy/integrand/linear0.py
@@ -8,10 +8,10 @@ class Linear0(Integrand):
     >>> l = Linear0(DigitalNetB2(100,seed=7))
     >>> x = l.discrete_distrib.gen_samples(2**10)
     >>> y = l.f(x)
-    >>> y.mean()
+    >>> float(y.mean())
     -1.175...e-08
     >>> ytf = l.f(x,periodization_transform='C1SIN')
-    >>> ytf.mean()
+    >>> float(ytf.mean())
     -4.050...e-12
     """
